Maintenance

+++ source/core/image/texture/sampler/sampler_2d_linear.inl
@@ -8,8 +8,7 @@ namespace image { namespace texture { namespace sampler {
 
 inline float bilinear_1(float c00, float c01, float c10, float c11, float s, float t);
 
-inline math::float2 bilinear_2(math::float2 c00, math::float2 c01,
-							   math::float2 c10, math::float2 c11,
+inline math::float2 bilinear_2(math::float2 c00, math::float2 c01, math::float2 c10, math::float2 c11,
 							   float s, float t);
 
 template<typename T>
@@ -28,11 +27,11 @@ float Sampler_2D_linear<Address_mode>::sample_1(const Texture_2D& texture, math:
 	float fu = std::floor(u);
 	float fv = std::floor(v);
 
-	int x = static_cast<int>(fu);
-	int y = static_cast<int>(fv);
+	int32_t x = static_cast<int32_t>(fu);
+	int32_t y = static_cast<int32_t>(fv);
 
-	int x1 = std::min(x + 1, static_cast<int>(d.x) - 1);
-	int y1 = std::min(y + 1, static_cast<int>(d.y) - 1);
+	int32_t x1 = std::min(x + 1, d.x - 1);
+	int32_t y1 = std::min(y + 1, d.y - 1);
 
 	x = std::max(x, 0);
 	y = std::max(y, 0);
@@ -61,11 +60,11 @@ math::float2 Sampler_2D_linear<Address_mode>::sample_2(const Texture_2D& texture
 	float fu = std::floor(u);
 	float fv = std::floor(v);
 
-	int x = static_cast<int>(fu);
-	int y = static_cast<int>(fv);
+	int32_t x = static_cast<int32_t>(fu);
+	int32_t y = static_cast<int32_t>(fv);
 
-	int x1 = std::min(x + 1, static_cast<int>(d.x) - 1);
-	int y1 = std::min(y + 1, static_cast<int>(d.y) - 1);
+	int x1 = std::min(x + 1, d.x - 1);
+	int y1 = std::min(y + 1, d.y - 1);
 
 	x = std::max(x, 0);
 	y = std::max(y, 0);
@@ -94,11 +93,11 @@ math::float3 Sampler_2D_linear<Address_mode>::sample_3(const Texture_2D& texture
 	float fu = std::floor(u);
 	float fv = std::floor(v);
 
-	int x = static_cast<int>(fu);
-	int y = static_cast<int>(fv);
+	int32_t x = static_cast<int32_t>(fu);
+	int32_t y = static_cast<int32_t>(fv);
 
-	int x1 = std::min(x + 1, static_cast<int>(d.x) - 1);
-	int y1 = std::min(y + 1, static_cast<int>(d.y) - 1);
+	int x1 = std::min(x + 1, d.x - 1);
+	int y1 = std::min(y + 1, d.y - 1);
 
 	x = std::max(x, 0);
 	y = std::max(y, 0);
@@ -126,8 +125,7 @@ inline float bilinear_1(float c00, float c01, float c10, float c11, float s, flo
 	return _s * (_t * c00 + t * c01) + s * (_t * c10 + t * c11);
 }
 
-inline math::float2 bilinear_2(math::float2 c00, math::float2 c01,
-							   math::float2 c10, math::float2 c11,
+inline math::float2 bilinear_2(math::float2 c00, math::float2 c01, math::float2 c10, math::float2 c11,
 							   float s, float t) {
 	float _s = 1.f - s;
 	float _t = 1.f - t;
